(PC-23190)[API] chore: extract post product test
